create table invoice compoent
